pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "werkzeug>=3.1.3",
    "psycopg2-binary>=2.9.10",
    "email-validator>=2.2.0",
    "gunicorn>=23.0.0",
    "python-pptx>=1.0.2",
    "docx2pdf>=0.1.8",
    "pypandoc>=1.15",
]
